First draw!
In this commit, I have been able to draw the canvas and player dots.

This is done by having the canvas prototype being the main parent
prototype and players are spawned off of the canvas methods.

My next step probably would be to have the player be controled my
keyboard inputs.