Organization and renderer orientation.
First I extracted the renderer as a mixin instead of a prototype.  Makes 
more sense this way. 

Player now spawns in the middle.  Default spawn spawn could be a 
setting?

I was able to add in a renderer orientation to the screen.  It worked 
suprisingly well for the moment.  Not much jankiness.